GH-7: add active boolean and triggers for messages
This update adds the active boolean column as well as triggers to
update the active_at and active columns when active is set/unset.

This update is to take logic out of the routes for setting the
active_at column and moves them to the database where they won't
be missed.